Restructure project completely
* Some special folders were created for scalability.
* Styles become modules and now got imported not from global but on places
* Tests are co-located for readability and usability
