+++ src/app/router.tsx
@@ -1,6 +1,6 @@
 import { Navigate, Route, Routes } from 'react-router-dom';
-import NotFound from '@/pages/NotFound';
-import CocktailPage from '@/pages/CocktailPage';
+import NotFound from '@/pages/NotFoundPage/NotFound';
+import CocktailPage from '@/pages/CocktailPage/CocktailPage';
 import { VALID_CODES } from '@/constants';
 
 export const AppRoutes = () => (
