+++ src/entities/drink/model/useCocktailStore.ts
@@ -1,15 +1,5 @@
+import { Drink } from '@/entities/drink/types';
 import { create } from 'zustand';
-
-export interface Drink {
-  idDrink: string;
-  strDrink: string;
-  strCategory: string;
-  strAlcoholic: string;
-  strGlass: string;
-  strInstructions: string;
-  strDrinkThumb: string;
-  [key: string]: string | null;
-}
 
 interface CocktailStore {
   data: Record<string, Drink[]>;
