+++ src/entities/drink/types.ts
@@ -1,0 +1,10 @@
+export interface Drink {
+  idDrink: string;
+  strDrink: string;
+  strCategory: string;
+  strAlcoholic: string;
+  strGlass: string;
+  strInstructions: string;
+  strDrinkThumb: string;
+  [key: string]: string | null;
+}
