+++ src/entities/drink/ui/DrinkCard/DrinkCard.module.scss
