+++ src/main.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from '@/app/App';
 import { BrowserRouter } from 'react-router-dom';
-import '@/shared/styles/main.scss';
+import '@/shared/styles/globals.scss';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
