+++ src/pages/CocktailPage/CocktailPage.tsx
@@ -1,10 +1,10 @@
 import { Navigate, useParams } from 'react-router-dom';
 import { useEffect, useMemo } from 'react';
-import DrinkCard from '@/shared/components/DrinkCard';
+import DrinkCard from '@/entities/drink/ui/DrinkCard/DrinkCard';
 import { VALID_CODES } from '@/constants';
 import throttle from 'lodash.throttle';
-import ErrorMessage from '@/shared/components/ErrorMessage';
-import { useFetchCocktails } from '@/hooks/useFetchCocktails';
+import { useFetchCocktails } from '@/shared/hooks/useFetchCocktails';
+import ErrorMessage from '@/shared/ui/ErrorMessage/ErrorMessage';
 
 export default function CocktailPage() {
   const { code } = useParams();
