+++ src/shared/styles/globals.scss
@@ -1,8 +1,3 @@
-@use './DrinkCard' as *;
-@use './ErrorMessage' as *;
-@use './NotFound' as *;
-@use './Sidebar' as *;
-
 body {
   margin: 0;
   font-family: sans-serif;
