+++ src/shared/ui/ErrorMessage/ErrorMessage.tsx
@@ -1,3 +1,5 @@
+import styles from './ErrorMessage.module.scss';
+
 interface ErrorMessageProps {
   error: string;
   onRetry: () => void;
@@ -5,9 +7,9 @@ interface ErrorMessageProps {
 
 export default function ErrorMessage({ error, onRetry }: ErrorMessageProps) {
   return (
-    <div className="error">
-      <p className="error__message">{error}</p>
-      <button className="error__button" onClick={onRetry}>
+    <div className={styles.error}>
+      <p className={styles.error__message}>{error}</p>
+      <button className={styles.error__button} onClick={onRetry}>
         Retry the request
       </button>
     </div>
