+++ src/shared/ui/Sidebar/Sidebar.module.scss
