+++ src/shared/ui/Sidebar/Sidebar.tsx
@@ -1,13 +1,14 @@
 import { NavLink } from 'react-router-dom';
+import styles from './Sidebar.module.scss';
 
 export default function Sidebar({ codes }: { codes: string[] }) {
   return (
-    <aside className="sidebar">
+    <aside className={styles.sidebar}>
       {codes.map((code) => (
         <NavLink
           key={code}
           to={`/${code}`}
-          className={({ isActive }) => (isActive ? 'active' : '')}
+          className={({ isActive }) => (isActive ? styles.active : '')}
         >
           {code[0].toUpperCase() + code.slice(1)}
         </NavLink>
